Removed netCDF4 package from integration tests of fixes
