+++ tests/integration/cmor/_fixes/cmip6/test_cnrm_cm6_1.py
@@ -4,81 +4,16 @@ import os
 import iris
 import numpy as np
 import pytest
-from netCDF4 import Dataset
 
 from esmvalcore.cmor._fixes.cmip6.cnrm_cm6_1 import Cl, Clcalipso, Cli, Clw
 from esmvalcore.cmor.fix import Fix
 from esmvalcore.cmor.table import get_var_info
 
+SAMPLE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
+                           os.pardir, 'sample_data')
+
 
 @pytest.fixture
-def cl_file(tmp_path):
-    """Create netcdf file with similar issues as ``cl``."""
-    nc_path = os.path.join(tmp_path, 'cnrm_cm6_1_cl.nc')
-    dataset = Dataset(nc_path, mode='w')
-    dataset.createDimension('time', size=1)
-    dataset.createDimension('lev', size=3)
-    dataset.createDimension('lat', size=2)
-    dataset.createDimension('lon', size=2)
-    dataset.createDimension('bnds', size=2)
-
-    # Dimensional variables
-    dataset.createVariable('time', np.float64, dimensions=('time',))
-    dataset.createVariable('lev', np.float64, dimensions=('lev',))
-    dataset.createVariable('lev_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.createVariable('lat', np.float64, dimensions=('lat',))
-    dataset.createVariable('lon', np.float64, dimensions=('lon',))
-    dataset.variables['time'][:] = [0.0]
-    dataset.variables['time'].standard_name = 'time'
-    dataset.variables['time'].units = 'days since 6543-2-1'
-    dataset.variables['lev'][:] = [1.0, 2.0, 4.0]
-    dataset.variables['lev'].standard_name = (
-        'atmosphere_hybrid_sigma_pressure_coordinate')
-    dataset.variables['lev'].bounds = 'lev_bnds'
-    dataset.variables['lev'].units = '1'
-    dataset.variables['lev'].formula_term = (
-        'ap: ap b: b ps: ps')  # Error in attribute intended
-    dataset.variables['lev_bnds'][:] = [[0.5, 1.5], [1.5, 3.0], [3.0, 5.0]]
-    dataset.variables['lev_bnds'].standard_name = (
-        'atmosphere_hybrid_sigma_pressure_coordinate')
-    dataset.variables['lev_bnds'].units = '1'
-    dataset.variables['lev_bnds'].formula_term = (
-        'ap: ap b: b ps: ps')  # Error in attribute intended
-    dataset.variables['lat'][:] = [-30.0, 0.0]
-    dataset.variables['lat'].standard_name = 'latitude'
-    dataset.variables['lat'].units = 'degrees_north'
-    dataset.variables['lon'][:] = [30.0, 60.0]
-    dataset.variables['lon'].standard_name = 'longitude'
-    dataset.variables['lon'].units = 'degrees_east'
-
-    # Coordinates for derivation of pressure coordinate
-    # Wrong shape of bounds is intended
-    dataset.createVariable('ap', np.float64, dimensions=('lev',))
-    dataset.createVariable('ap_bnds', np.float64, dimensions=('bnds', 'lev'))
-    dataset.createVariable('b', np.float64, dimensions=('lev',))
-    dataset.createVariable('b_bnds', np.float64, dimensions=('bnds', 'lev'))
-    dataset.createVariable('ps', np.float64,
-                           dimensions=('time', 'lat', 'lon'))
-    dataset.variables['ap'][:] = [1.0, 2.0, 5.0]
-    dataset.variables['ap_bnds'][:] = [[0.0, 1.5, 1.5], [3.0, 3.0, 6.0]]
-    dataset.variables['b'][:] = [0.0, 1.0, 3.0]
-    dataset.variables['b_bnds'][:] = [[-1.0, 0.5, 0.5], [2.0, 2.0, 5.0]]
-    dataset.variables['ps'][:] = np.arange(1 * 2 * 2).reshape(1, 2, 2)
-    dataset.variables['ps'].standard_name = 'surface_air_pressure'
-    dataset.variables['ps'].units = 'Pa'
-
-    # Cl variable
-    dataset.createVariable('cl', np.float32,
-                           dimensions=('time', 'lev', 'lat', 'lon'))
-    dataset.variables['cl'][:] = np.full((1, 3, 2, 2), 0.0, dtype=np.float32)
-    dataset.variables['cl'].standard_name = (
-        'cloud_area_fraction_in_atmosphere_layer')
-    dataset.variables['cl'].units = '%'
-
-    dataset.close()
-    return nc_path
-
-
 def test_get_cl_fix():
     """Test getting of fix."""
     fix = Fix.get_fixes('CMIP6', 'CNRM-CM6-1', 'Amon', 'cl')
@@ -105,9 +40,10 @@ AIR_PRESSURE_BOUNDS = np.array([[[[[0.0, 1.5],
                                    [9.0, 21.0]]]]])
 
 
-def test_cl_fix_metadata(cl_file):
+def test_cl_fix_metadata():
     """Test ``fix_metadata`` for ``cl``."""
-    cubes = iris.load(cl_file)
+    nc_path = os.path.join(SAMPLE_PATH, 'cnrm_cm6_1_cl.nc')
+    cubes = iris.load(nc_path)
 
     # Raw cubes
     assert len(cubes) == 6
