+++ tests/integration/cmor/_fixes/test_common.py
@@ -3,156 +3,16 @@ import os
 
 import iris
 import numpy as np
-import pytest
-from netCDF4 import Dataset
 
-from esmvalcore.cmor._fixes.common import (ClFixHybridHeightCoord,
-                                           ClFixHybridPressureCoord)
+from esmvalcore.cmor._fixes.common import (
+    ClFixHybridHeightCoord,
+    ClFixHybridPressureCoord,
+)
 from esmvalcore.cmor.table import get_var_info
 from esmvalcore.iris_helpers import var_name_constraint
 
-
-def create_hybrid_pressure_file_without_ap(dataset, short_name):
-    """Create dataset without vertical auxiliary coordinate ``ap``."""
-    dataset.createDimension('time', size=1)
-    dataset.createDimension('lev', size=2)
-    dataset.createDimension('lat', size=3)
-    dataset.createDimension('lon', size=4)
-    dataset.createDimension('bnds', size=2)
-
-    # Dimensional variables
-    dataset.createVariable('time', np.float64, dimensions=('time',))
-    dataset.createVariable('lev', np.float64, dimensions=('lev',))
-    dataset.createVariable('lev_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.createVariable('lat', np.float64, dimensions=('lat',))
-    dataset.createVariable('lon', np.float64, dimensions=('lon',))
-    dataset.variables['time'][:] = [0.0]
-    dataset.variables['time'].standard_name = 'time'
-    dataset.variables['time'].units = 'days since 6543-2-1'
-    dataset.variables['lev'][:] = [1.0, 2.0]
-    dataset.variables['lev'].bounds = 'lev_bnds'
-    dataset.variables['lev'].standard_name = (
-        'atmosphere_hybrid_sigma_pressure_coordinate')
-    dataset.variables['lev'].units = '1'
-    dataset.variables['lev_bnds'][:] = [[0.5, 1.5], [1.5, 3.0]]
-    dataset.variables['lev_bnds'].standard_name = (
-        'atmosphere_hybrid_sigma_pressure_coordinate')
-    dataset.variables['lev_bnds'].units = '1'
-    dataset.variables['lat'][:] = [-30.0, 0.0, 30.0]
-    dataset.variables['lat'].standard_name = 'latitude'
-    dataset.variables['lat'].units = 'degrees_north'
-    dataset.variables['lon'][:] = [30.0, 60.0, 90.0, 120.0]
-    dataset.variables['lon'].standard_name = 'longitude'
-    dataset.variables['lon'].units = 'degrees_east'
-
-    # Coordinates for derivation of pressure coordinate
-    dataset.createVariable('b', np.float64, dimensions=('lev',))
-    dataset.createVariable('b_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.createVariable('ps', np.float64,
-                           dimensions=('time', 'lat', 'lon'))
-    dataset.variables['b'][:] = [0.0, 1.0]
-    dataset.variables['b_bnds'][:] = [[-1.0, 0.5], [0.5, 2.0]]
-    dataset.variables['ps'][:] = np.arange(1 * 3 * 4).reshape(1, 3, 4)
-    dataset.variables['ps'].standard_name = 'surface_air_pressure'
-    dataset.variables['ps'].units = 'Pa'
-    dataset.variables['ps'].additional_attribute = 'xyz'
-
-    # Variable
-    dataset.createVariable(short_name, np.float32,
-                           dimensions=('time', 'lev', 'lat', 'lon'))
-    dataset.variables[short_name][:] = np.full((1, 2, 3, 4), 0.0,
-                                               dtype=np.float32)
-    dataset.variables[short_name].standard_name = (
-        'cloud_area_fraction_in_atmosphere_layer')
-    dataset.variables[short_name].units = '%'
-
-
-def create_hybrid_pressure_file_with_a(dataset, short_name):
-    """Create netcdf file with issues in hybrid pressure coordinate."""
-    create_hybrid_pressure_file_without_ap(dataset, short_name)
-    dataset.createVariable('a', np.float64, dimensions=('lev',))
-    dataset.createVariable('a_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.createVariable('p0', np.float64, dimensions=())
-    dataset.variables['a'][:] = [1.0, 2.0]
-    dataset.variables['a_bnds'][:] = [[0.0, 1.5], [1.5, 3.0]]
-    dataset.variables['p0'][:] = 1.0
-    dataset.variables['p0'].units = 'Pa'
-    dataset.variables['lev'].formula_terms = 'p0: p0 a: a b: b ps: ps'
-    dataset.variables['lev_bnds'].formula_terms = (
-        'p0: p0 a: a_bnds b: b_bnds ps: ps')
-
-
-def create_hybrid_pressure_file_with_ap(dataset, short_name):
-    """Create netcdf file with issues in hybrid pressure coordinate."""
-    create_hybrid_pressure_file_without_ap(dataset, short_name)
-    dataset.createVariable('ap', np.float64, dimensions=('lev',))
-    dataset.createVariable('ap_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.variables['ap'][:] = [1.0, 2.0]
-    dataset.variables['ap_bnds'][:] = [[0.0, 1.5], [1.5, 3.0]]
-    dataset.variables['ap'].units = 'Pa'
-    dataset.variables['lev'].formula_terms = 'ap: ap b: b ps: ps'
-    dataset.variables['lev_bnds'].formula_terms = (
-        'ap: ap_bnds b: b_bnds ps: ps')
-
-
-@pytest.fixture
-def cl_file_with_a(tmp_path):
-    """Create netcdf file for ``cl`` with ``a`` coordinate."""
-    nc_path = os.path.join(tmp_path, 'cl_a.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_pressure_file_with_a(dataset, 'cl')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def cl_file_with_ap(tmp_path):
-    """Create netcdf file for ``cl`` with ``ap`` coordinate."""
-    nc_path = os.path.join(tmp_path, 'cl_ap.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_pressure_file_with_ap(dataset, 'cl')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def cli_file_with_a(tmp_path):
-    """Create netcdf file for ``cli`` with ``a`` coordinate."""
-    nc_path = os.path.join(tmp_path, 'cli_a.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_pressure_file_with_a(dataset, 'cli')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def cli_file_with_ap(tmp_path):
-    """Create netcdf file for ``cli`` with ``ap`` coordinate."""
-    nc_path = os.path.join(tmp_path, 'cli_ap.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_pressure_file_with_ap(dataset, 'cli')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def clw_file_with_a(tmp_path):
-    """Create netcdf file for ``clw`` with ``a`` coordinate."""
-    nc_path = os.path.join(tmp_path, 'clw_a.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_pressure_file_with_a(dataset, 'clw')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def clw_file_with_ap(tmp_path):
-    """Create netcdf file for ``clw`` with ``ap`` coordinate."""
-    nc_path = os.path.join(tmp_path, 'clw_ap.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_pressure_file_with_ap(dataset, 'clw')
-    dataset.close()
-    return nc_path
+SAMPLE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
+                           'sample_data')
 
 
 AIR_PRESSURE_POINTS = np.array([[[[1.0, 1.0, 1.0, 1.0],
@@ -187,8 +47,9 @@ AIR_PRESSURE_BOUNDS = np.array([[[[[0.0, 1.5],
                                    [7.0, 25.0]]]]])
 
 
-def hybrid_pressure_coord_fix_metadata(nc_path, short_name, fix):
+def hybrid_pressure_coord_fix_metadata(nc_filename, short_name, fix):
     """Test ``fix_metadata`` of file with hybrid pressure coord."""
+    nc_path = os.path.join(SAMPLE_PATH, nc_filename)
     cubes = iris.load(nc_path)
 
     # Raw cubes
@@ -226,106 +87,20 @@ def hybrid_pressure_coord_fix_metadata(nc_path, short_name, fix):
     return var_names
 
 
-def test_cl_hybrid_pressure_coord_fix_metadata_with_a(cl_file_with_a):
+def test_cl_hybrid_pressure_coord_fix_metadata_with_a():
     """Test ``fix_metadata`` for ``cl``."""
     vardef = get_var_info('CMIP6', 'Amon', 'cl')
     var_names = hybrid_pressure_coord_fix_metadata(
-        cl_file_with_a, 'cl', ClFixHybridPressureCoord(vardef))
+        'common_cl_a.nc', 'cl', ClFixHybridPressureCoord(vardef))
     assert 'a_bnds' in var_names
 
 
-def test_cl_hybrid_pressure_coord_fix_metadata_with_ap(cl_file_with_ap):
+def test_cl_hybrid_pressure_coord_fix_metadata_with_ap():
     """Test ``fix_metadata`` for ``cl``."""
     vardef = get_var_info('CMIP6', 'Amon', 'cl')
     var_names = hybrid_pressure_coord_fix_metadata(
-        cl_file_with_ap, 'cl', ClFixHybridPressureCoord(vardef))
+        'common_cl_ap.nc', 'cl', ClFixHybridPressureCoord(vardef))
     assert 'ap_bnds' in var_names
-
-
-def create_hybrid_height_file(dataset, short_name):
-    """Create dataset with hybrid height coordinate."""
-    dataset.createDimension('time', size=1)
-    dataset.createDimension('lev', size=2)
-    dataset.createDimension('lat', size=1)
-    dataset.createDimension('lon', size=2)
-    dataset.createDimension('bnds', size=2)
-
-    # Dimensional variables
-    dataset.createVariable('time', np.float64, dimensions=('time',))
-    dataset.createVariable('lev', np.float64, dimensions=('lev',))
-    dataset.createVariable('lev_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.createVariable('lat', np.float64, dimensions=('lat',))
-    dataset.createVariable('lon', np.float64, dimensions=('lon',))
-    dataset.variables['time'][:] = [0.0]
-    dataset.variables['time'].standard_name = 'time'
-    dataset.variables['time'].units = 'days since 6543-2-1'
-    dataset.variables['lev'][:] = [1.0, 2.0]
-    dataset.variables['lev'].bounds = 'lev_bnds'
-    dataset.variables['lev'].standard_name = (
-        'atmosphere_hybrid_height_coordinate')
-    dataset.variables['lev'].units = 'm'
-    dataset.variables['lev'].formula_terms = 'a: lev b: b orog: orog'
-    dataset.variables['lev_bnds'][:] = [[0.5, 1.5], [1.5, 3.0]]
-    dataset.variables['lev_bnds'].standard_name = (
-        'atmosphere_hybrid_height_coordinate')
-    dataset.variables['lev_bnds'].units = '1'
-    dataset.variables['lev_bnds'].formula_terms = (
-        'a: lev_bnds b: b_bnds orog: orog')
-    dataset.variables['lat'][:] = [0.0]
-    dataset.variables['lat'].standard_name = 'latitude'
-    dataset.variables['lat'].units = 'degrees_north'
-    dataset.variables['lon'][:] = [30.0, 60.0]
-    dataset.variables['lon'].standard_name = 'longitude'
-    dataset.variables['lon'].units = 'degrees_east'
-
-    # Coordinates for derivation of height coordinate
-    dataset.createVariable('b', np.float64, dimensions=('lev',))
-    dataset.createVariable('b_bnds', np.float64, dimensions=('lev', 'bnds'))
-    dataset.createVariable('orog', np.float64, dimensions=('lat', 'lon'))
-    dataset.variables['b'][:] = [0.0, 1.0]
-    dataset.variables['b_bnds'][:] = [[-1.0, 0.5], [0.5, 2.0]]
-    dataset.variables['orog'][:] = [[0.0, 1.0]]
-    dataset.variables['orog'].standard_name = 'surface_altitude'
-    dataset.variables['orog'].units = 'm'
-
-    # Variable
-    dataset.createVariable(short_name, np.float32,
-                           dimensions=('time', 'lev', 'lat', 'lon'))
-    dataset.variables[short_name][:] = np.full((1, 2, 1, 2), 0.0,
-                                               dtype=np.float32)
-    dataset.variables[short_name].standard_name = (
-        'cloud_area_fraction_in_atmosphere_layer')
-    dataset.variables[short_name].units = '%'
-
-
-@pytest.fixture
-def cl_file_with_height(tmp_path):
-    """Create netcdf file for ``cl`` with hybrid height coordinate."""
-    nc_path = os.path.join(tmp_path, 'cl_hybrid_height.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_height_file(dataset, 'cl')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def cli_file_with_height(tmp_path):
-    """Create netcdf file for ``cli`` with hybrid height coordinate."""
-    nc_path = os.path.join(tmp_path, 'cli_hybrid_height.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_height_file(dataset, 'cli')
-    dataset.close()
-    return nc_path
-
-
-@pytest.fixture
-def clw_file_with_height(tmp_path):
-    """Create netcdf file for ``clw`` with hybrid height coordinate."""
-    nc_path = os.path.join(tmp_path, 'clw_hybrid_height.nc')
-    dataset = Dataset(nc_path, mode='w')
-    create_hybrid_height_file(dataset, 'clw')
-    dataset.close()
-    return nc_path
 
 
 HEIGHT_POINTS = np.array([[[1.0, 1.0]],
@@ -346,8 +121,9 @@ PRESSURE_BOUNDS = np.array([[[[101318.99243691, 101306.9780559],
                               [101300.97123885, 101264.93559234]]]])
 
 
-def hybrid_height_coord_fix_metadata(nc_path, short_name, fix):
+def hybrid_height_coord_fix_metadata(nc_filename, short_name, fix):
     """Test ``fix_metadata`` of file with hybrid height coord."""
+    nc_path = os.path.join(SAMPLE_PATH, nc_filename)
     cubes = iris.load(nc_path)
 
     # Raw cubes
@@ -386,8 +162,8 @@ def hybrid_height_coord_fix_metadata(nc_path, short_name, fix):
     assert air_pressure_coord.units == 'Pa'
 
 
-def test_cl_hybrid_height_coord_fix_metadata(cl_file_with_height):
+def test_cl_hybrid_height_coord_fix_metadata():
     """Test ``fix_metadata`` for ``cl``."""
     vardef = get_var_info('CMIP6', 'Amon', 'cl')
-    hybrid_height_coord_fix_metadata(cl_file_with_height, 'cl',
+    hybrid_height_coord_fix_metadata('common_cl_hybrid_height.nc', 'cl',
                                      ClFixHybridHeightCoord(vardef))
